+++ plugins/polio/js/src/domains/Campaigns/MapComponent/MapComponent.js
@@ -95,6 +95,7 @@ export const MapComponent = ({
             </Pane>
             <PaneWithPattern
                 name={`MainLayer-${name}`}
+                key={name}
                 patterns={shapePatterns}
                 patternIds={shapePatternIds}
             >
