+++ plugins/polio/js/src/domains/LQAS-IM/shared/DebugInfo/BadRoundNumber.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react';
 import { useSafeIntl } from 'bluesquare-components';
 import { Typography } from '@mui/material';
-import MESSAGES from '../../../constants/messages';
+import MESSAGES from '../../../../constants/messages';
 
 type Props = {
     // eslint-disable-next-line react/require-default-props
