+++ plugins/polio/js/src/domains/LQAS-IM/shared/DebugInfo/DistrictsNotFound.tsx
@@ -1,8 +1,8 @@
 import React, { FunctionComponent } from 'react';
 import { useSafeIntl } from 'bluesquare-components';
 import { Typography, Chip, Box } from '@mui/material';
-import { LqasImCampaign } from '../../../constants/types';
-import MESSAGES from '../../../constants/messages';
+import { LqasImCampaign } from '../../../../constants/types';
+import MESSAGES from '../../../../constants/messages';
 
 type Props = {
     // eslint-disable-next-line react/require-default-props
