+++ plugins/polio/js/src/domains/LQAS-IM/shared/Map/LqasImMapHeader.tsx
@@ -3,10 +3,10 @@ import { Box, Divider, Grid, Typography } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import { useSafeIntl } from 'bluesquare-components';
 import classNames from 'classnames';
-import MESSAGES from '../../../constants/messages';
+import MESSAGES from '../../../../constants/messages';
 import { LqasImDates } from './LqasImDates';
-import { DropdownOptions } from '../../../../../../../hat/assets/js/apps/Iaso/types/utils';
-import InputComponent from '../../../../../../../hat/assets/js/apps/Iaso/components/forms/InputComponent';
+import { DropdownOptions } from '../../../../../../../../hat/assets/js/apps/Iaso/types/utils';
+import InputComponent from '../../../../../../../../hat/assets/js/apps/Iaso/components/forms/InputComponent';
 
 export type LqasImRefDate = {
     date: string;
