+++ plugins/polio/js/src/domains/LQAS-IM/shared/Map/ScopeAndDNFDisclaimer.tsx
@@ -5,8 +5,8 @@ import { makeStyles } from '@mui/styles';
 import React, { FunctionComponent } from 'react';
 // @ts-ignore
 import { useSafeIntl } from 'bluesquare-components';
-import MESSAGES from '../../../constants/messages';
-import { findScopeIds } from '../../../utils';
+import MESSAGES from '../../../../constants/messages';
+import { findScopeIds } from '../../../../utils';
 
 type Props = {
     campaign?: string;
