+++ plugins/polio/js/src/domains/LQAS-IM/shared/charts/NoData.tsx
@@ -2,7 +2,7 @@ import { Box, Grid, Typography } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import React, { FunctionComponent } from 'react';
 import { useSafeIntl } from 'bluesquare-components';
-import MESSAGES from '../../../constants/messages';
+import MESSAGES from '../../../../constants/messages';
 
 const style = theme => ({
     paper: {
