+++ plugins/polio/js/src/domains/LQAS-IM/shared/charts/PercentageBarChart/constants.js
