+++ plugins/polio/js/src/domains/LQAS-IM/shared/charts/PercentageChartWithTitle.tsx
