+++ plugins/polio/js/src/domains/LQAS-IM/shared/hooks/api/useLqasIm.ts
@@ -8,9 +8,9 @@ import {
     IM_GLOBAL_SLUG,
     IM_OHH_SLUG,
     IM_HH_SLUG,
-} from '../IM/constants';
-import { LqasImData } from '../../../constants/types';
-import { LQAS_COUNTRY_URL } from '../LQAS/constants';
+} from '../../../IM/constants';
+import { LqasImData } from '../../../../../constants/types';
+import { LQAS_COUNTRY_URL } from '../../../LQAS/constants';
 
 export type LQASIMRequestType = 'lqas' | 'imOHH' | 'imIHH' | 'imGlobal';
